Create a Tkinter window for program output
A Frame was created in Tkinter and within it two widget text boxes were placed, side by side, one for the original text and the other for the translated text. The name of the folder where the files were saved and the processing time are also shown, for those options where this information is required.