Remove the timeout parameter and include the phrase_time_limit parameter
Remove the timeout parameter and include the phrase_time_limit=5 parameter. The phrase_time_limit parameter is the maximum number of seconds that this will allow a phrase to continue. Whereas, The timeout parameter is the maximum number of seconds that this will wait for a phrase to start. So, as we want to limit the time for the microphone to stop listening, it was necessary to add phrase_time_limit.